+++ Shim/ChaincodeServer.cs
@@ -9,10 +9,21 @@ namespace Shim {
     /// The ChaincodeServer class represents a chaincode gRPC server, which waits for connections from peers.
     /// </summary>
     public class ChaincodeServer: Chaincode.ChaincodeBase{
-        public string CCID { get; } // CCID should match chaincode's package name on peer
-        public string Address { get; }// Addresss is the listen address of the chaincode server
 
-        public IChaincode Chaincode { get; set; }// CC is the chaincode that handles Init and Invoke
+        /// <summary>
+        ///  CCID should match chaincode's package name on peer
+        /// </summary>
+        public string CCID { get; }
+
+        /// <summary>
+        /// Addresss is the listen address of the chaincode server
+        /// </summary>
+        public string Address { get; }
+
+        /// <summary>
+        ///  CC is the chaincode that handles Init and Invoke
+        /// </summary>
+        public IChaincode Chaincode { get; set; }
 
         public ChaincodeServer(string chaincodeId , string address, IChaincode chaincode) {
             CCID = chaincodeId;
